requirements.txt
<<<<<<< HEAD
# Python Requirements (for reference/college project structure)
flask==3.0.0
flask_sqlalchemy==3.1.1
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
nltk==3.8.1
spacy==3.7.2
sumy==0.11.0
reportlab==4.0.7
=======
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.2.0
nltk==3.8.1
sumy==0.11.0
reportlab==4.0.8
Werkzeug==3.0.1
pdf2image==1.17.0
>>>>>>> 58b7d10 (new idea)
